requirements.txt
matplotlib
opencv-python
Pillow
requests
scikit-image
scikit-learn
scipy
webcolors
Werkzeug
bs4
click
anyio
certifi
exceptiongroup
h11
httpcore
httpx
idna
python-telegram-bot[job-queue]
sniffio
typing_extensions
pyyaml
coloredlogs
slugify
pathvalidate
cyrtranslit
telebot